Add support for categories
Main page now shows categories for all blog postings. Thanks to Spencer
for the help with the CSS.
A click on a category lists all postings in this category.
A single posting lists all categories below the headline.

I'm unhappy with the hack in themes/pivotal-ui/layouts/partials/head.html
and themes/pivotal-ui/layouts/partials/heading.html, but I couldn't find
another way to extract the original title of a category, if the category
get's mangled by "urlize" and things like an ampersend are removed.