Git repo moved from _pivotal_ org to _pivotal-legacy_ org
The blog hasn't published for two weeks, and I suspect the problem is
the GitHub org-change. This is an attempt at remedy.

fixes <https://travis-ci.org/pivotal-legacy/blog>:
```
Skipping a deployment with the cloudfoundry provider because this repo's name does not match one specified in .travis.yml's deploy.on.repo: pivotal/blog
```