Troubleshooting obscure OpenSSH failures
By using tcpdump to troubleshoot an elusive error, we uncovered a
man-in-the-middle (MITM) ssh proxy installed by our information security
(InfoSec) team to harden/protect a set of machines which were accessible
from the internet. The ssh proxy in question was Palo Alto Network’s
(PAN) Layer 7 (i.e. it worked on any port, not solely ssh’s port 22)
proxy, and was discovered when we observed a failure to negotiate
ciphers during the ssh key exchange.

Signed-off-by: Brian Cunnie <bcunnie@pivotal.io>